# backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: Leave Management API
  version: 2.0.0
  description: MERN Leave Management Backend (Trash-only flow)
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, example: aljun.dalman@neu.edu.ph }
                password: { type: string, example: Admin123 }
      responses:
        '200': { description: Successful login }
        '401': { description: Invalid credentials }
  /auth/me:
    get:
      summary: Get current user profile
      tags: [Auth]
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: Current user details }
        '404': { description: User not found }
  /auth/change-password:
    post:
      summary: Change password
      tags: [Auth]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200': { description: Password changed successfully }
        '400': { description: Current password incorrect }
  /auth/logout:
    post:
      summary: Logout
      tags: [Auth]
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: Logout successful }
        '401': { description: Unauthorized – no token or invalid token }
...